Improve error reporting
BEFORE: If there is a group with a transform somewhere in the middle of
the input SVG, it yields the message "<g> with a transform not handled"
at the corresponding position in the vector icon output, where you might
not notice. If it is at the end, then the word handled becomes handle,
because of code intended to remove a comma from the end of the output.

AFTER: If there is a group with a transform in the input SVG, then
"Error: <g> with a transform not handled" is the whole output. Any
JavaScript error in HandleNode is reported in a similar fashion.